requirements.txt
beautifulsoup4==4.6.0
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
Flask==1.1.2
google-api-core==1.31.0
google-auth==1.33.1
google-cloud-vision==2.4.0
googleapis-common-protos==1.53.0
grpcio==1.38.1
idna==3.2
importlib-metadata==3.10.0
itsdangerous==2.0.1
Jinja2==3.0.1
JPype1==1.3.0
konlpy==0.5.2
lxml==4.6.3
MarkupSafe==2.0.1
numpy==1.21.1
oauthlib==3.1.1
packaging==21.0
pip==21.1.3
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
PySocks==1.7.1
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
setuptools==57.4.0
six==1.16.0
tweepy==3.10.0
urllib3==1.26.6
Werkzeug==1.0.1
wheel==0.36.2
wincertstore==0.2
zipp==3.5.0
